Enhancement: use the Design tab's color control for the embed button color
The external tab now renders PanelColorSettings with the same default
palette as the Design tab's Primary Color, so the two look and behave
alike. The palette moves to defaultColors.ts so both share it.
