+++ src/FormBuilder/resources/js/form-builder/src/settings/design/style-controls/color/defaultColors.ts
@@ -1,0 +1,22 @@
+/**
+ * The palette offered wherever the builder asks for a form color: the Design
+ * tab's primary and secondary colors and the external embed's button color.
+ *
+ * @since 4.3.0 Update the value of the default colors Primary color to improve accessibility color contrast.
+ * @since TBD Shared with the external embed button color.
+ */
+const defaultColors = [
+    {name: 'Black', slug: 'black', color: '#000000'},
+    {name: 'Dark Blue', slug: 'dark-blue', color: '#1E1AE2'},
+    {name: 'Give Primary Default', slug: 'give-primary-default', color: '#2d802f'},
+    {name: 'Red', slug: 'red', color: '#BD3D36'},
+    {name: 'Orange', slug: 'orange', color: '#EB712E'},
+    {name: 'Gray', slug: 'gray', color: '#5F7385'},
+    {name: 'Light Blue', slug: 'light-blue', color: '#4492DD'},
+    {name: 'Light Green', slug: 'light-green', color: '#63CC8A'},
+    {name: 'Purple', slug: 'purple', color: '#9058D8'},
+    {name: 'Teal', slug: 'teal', color: '#2BBAB1'},
+    {name: 'Yellow', slug: 'yellow', color: '#F1BB40'},
+];
+
+export default defaultColors;
